test: move create_shelf_modal tests to sibling tests.rs
